mv qt view actions
