+++ napari/_qt/_qapp_model/qactions/_view.py
@@ -1,7 +1,7 @@
 """Actions related to the 'View' menu that require Qt.
 
-'View' actions that do not require Qt should go in
-`napari/_app_model/actions/_view_actions.py`.
+'View' actions that do not require Qt should be placed under
+`napari/_app_model/actions/` in a `_view_actions.py` file.
 """
 
 import sys
@@ -15,9 +15,11 @@ from app_model.types import (
     ToggleRule,
 )
 
+from napari._app_model.actions._toggle_action import ViewerToggleAction
 from napari._app_model.constants import CommandId, MenuGroup, MenuId
 from napari._qt.qt_main_window import Window
 from napari._qt.qt_viewer import QtViewer
+from napari.settings import get_settings
 from napari.utils.translations import trans
 
 
@@ -39,6 +41,15 @@ def _get_current_play_status(qt_viewer: QtViewer):
 
 def _get_current_activity_dock_status(window: Window):
     return window._qt_window._activity_dialog.isVisible()
+
+
+def _tooltip_visibility_toggle() -> None:
+    settings = get_settings().appearance
+    settings.layer_tooltip_visibility = not settings.layer_tooltip_visibility
+
+
+def _get_current_tooltip_visibility() -> bool:
+    return get_settings().appearance.layer_tooltip_visibility
 
 
 Q_VIEW_ACTIONS: List[Action] = [
@@ -102,4 +113,41 @@ Q_VIEW_ACTIONS: List[Action] = [
         callback=_toggle_activity_dock,
         toggled=ToggleRule(get_current=_get_current_activity_dock_status),
     ),
+    # TODO: this could be made into a toggle setting Action subclass
+    # using a similar pattern to the above ViewerToggleAction classes
+    Action(
+        id=CommandId.TOGGLE_LAYER_TOOLTIPS,
+        title=CommandId.TOGGLE_LAYER_TOOLTIPS.command_title,
+        menus=[
+            {
+                'id': MenuId.MENUBAR_VIEW,
+                'group': MenuGroup.RENDER,
+                'order': 10,
+            }
+        ],
+        callback=_tooltip_visibility_toggle,
+        toggled=ToggleRule(get_current=_get_current_tooltip_visibility),
+    ),
 ]
+
+MENUID_DICT = {'axes': MenuId.VIEW_AXES, 'scale_bar': MenuId.VIEW_SCALEBAR}
+
+for cmd, viewer_attr, sub_attr in (
+    (CommandId.TOGGLE_VIEWER_AXES, 'axes', 'visible'),
+    (CommandId.TOGGLE_VIEWER_AXES_COLORED, 'axes', 'colored'),
+    (CommandId.TOGGLE_VIEWER_AXES_LABELS, 'axes', 'labels'),
+    (CommandId.TOGGLE_VIEWER_AXES_DASHED, 'axes', 'dashed'),
+    (CommandId.TOGGLE_VIEWER_AXES_ARROWS, 'axes', 'arrows'),
+    (CommandId.TOGGLE_VIEWER_SCALE_BAR, 'scale_bar', 'visible'),
+    (CommandId.TOGGLE_VIEWER_SCALE_BAR_COLORED, 'scale_bar', 'colored'),
+    (CommandId.TOGGLE_VIEWER_SCALE_BAR_TICKS, 'scale_bar', 'ticks'),
+):
+    Q_VIEW_ACTIONS.append(
+        ViewerToggleAction(
+            id=cmd,
+            title=cmd.command_title,
+            viewer_attribute=viewer_attr,
+            sub_attribute=sub_attr,
+            menus=[{'id': MENUID_DICT[viewer_attr]}],
+        )
+    )
